requirements.txt
appnope==0.1.4
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
directory_tree==1.0.0
executing==2.2.1
fonttools==4.61.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
urllib3==2.6.2
wcwidth==0.2.14
